Add a trait for using caches
Some providers may need a cache directory. The functions of the new
trait may be used to get a valid cache path inside of a global defined
cache path. This makes the configuration more simple, as only one cache
path needs to be defined but several providers may use it.